Move otel dependency handling to a separate package
